move failure reason tests to grz-db
